*: Add test case and fix bug
